#1860 - Updated API POST /{partnerId}/policies/{policyId}/bio-extractors-request with POST /partner-policy-requests/{requestId}/bio-extractors-request
