+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/controller/PartnerServiceController.java
@@ -91,9 +91,6 @@ public class PartnerServiceController {
 	@Value("${mosip.pms.api.id.partner.exists.post}")
 	private String postPartnerExistsId;
 
-	@Value("${mosip.pms.api.id.partners.bioextractors.request.post:mosip.pms.partners.bioextractors.request.post}")
-	private String postPartnerBioextractorsRequestId;
-
 	@Autowired
 	PartnerService partnerService;
 	
@@ -182,37 +179,6 @@ public class PartnerServiceController {
 		response.setId(request.getId());
 		response.setVersion(request.getVersion());
 		return new ResponseEntity<>(response, HttpStatus.OK);		
-	}
-
-	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
-	@PostMapping(value = "/partner-policy-requests/{requestId}/bio-extractors-request")
-	@Operation(summary = "Service to submit bio extractors request", description = "Persists bio extractor requests against an in-progress partner policy mapping request")
-	@ApiResponses(value = {
-			@ApiResponse(responseCode = "200", description = "OK"),
-			@ApiResponse(responseCode = "401", description = "Unauthorized", content = @Content(schema = @Schema(hidden = true))),
-			@ApiResponse(responseCode = "403", description = "Forbidden", content = @Content(schema = @Schema(hidden = true)))
-	})
-	public ResponseWrapperV2<String> submitBioExtractorsRequest(
-			@PathVariable("requestId") String requestId,
-			@RequestBody @Valid RequestWrapperV2<BioExtractorsRequestDto> requestWrapper) {
-		Optional<ResponseWrapperV2<String>> validationResponse =
-				requestValidator.validate(postPartnerBioextractorsRequestId, requestWrapper);
-		if (validationResponse.isPresent()) {
-			return validationResponse.get();
-		}
-		inputValidator.validateRequestInput("requestId", requestId);
-		requestWrapper.getRequest().getExtractors().forEach(extractor -> {
-			inputValidator.validateRequestInput("attributeName", extractor.getAttributeName());
-			inputValidator.validateRequestInput("biometric", extractor.getBiometric());
-			inputValidator.validateRequestInput("biometricSubTypes", extractor.getBiometricSubTypes());
-			inputValidator.validateRequestInput("extractorProvider", extractor.getExtractorProvider());
-			inputValidator.validateRequestInput("extractorProviderVersion", extractor.getExtractorProviderVersion());
-		});
-		ResponseWrapperV2<String> response = new ResponseWrapperV2<>();
-		response.setResponse(partnerService.submitBioExtractorsRequest(requestId, requestWrapper.getRequest()));
-		response.setId(requestWrapper.getId());
-		response.setVersion(requestWrapper.getVersion());
-		return response;
 	}
 
 	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
