+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/service/PartnerService.java
@@ -20,7 +20,6 @@ import io.mosip.pms.device.response.dto.FilterResponseCodeDto;
 import io.mosip.pms.partner.dto.*;
 import io.mosip.pms.partner.request.dto.AddContactRequestDto;
 import io.mosip.pms.partner.request.dto.ExtractorsDto;
-import io.mosip.pms.partner.request.dto.BioExtractorsRequestDto;
 import io.mosip.pms.partner.request.dto.CredentialTypeRequestDto;
 import io.mosip.pms.partner.request.dto.CACertificateRequestDto;
 import io.mosip.pms.partner.request.dto.PartnerPolicyMappingRequest;
@@ -138,15 +137,6 @@ public interface PartnerService {
      * @return
      */
     public String addBiometricExtractors(String partnerId, String policyId, ExtractorsDto extractors);
-
-	/**
-	 * Submit bio extractor configuration requests against an in-progress partner policy mapping request.
-	 *
-	 * @param requestId partner policy mapping request identifier
-	 * @param request bio-extractor request payload
-	 * @return created request row identifiers
-	 */
-	String submitBioExtractorsRequest(String requestId, BioExtractorsRequestDto request);
 
 	/**
 	 * Submit credential type request against an in-progress partner policy mapping request.
